pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "future-ralph"
version = "0.1.0"
description = "A heterogeneous agent wrapper exploring multiple possible futures."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "LightBytes", email = "contact@lightbytes.io"},
]
dependencies = [
  "typer>=0.9.0",
  "rich>=13.0.0",
  "pydantic>=2.0.0",
  "textual>=0.40.0",
  "pyyaml>=6.0",
]

[project.scripts]
future-ralph = "future_ralph.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/future_ralph"]
